chore/QB-1360 add local testing nodes
