+++ example/network/node1/configs/config.toml
@@ -1,5 +1,5 @@
 download_path_min_len = 88
-port = '9081'
+port = '9101'
 network_address = '127.0.0.1'
 debug = true
 pp_list_dir = "./peers"
@@ -20,11 +20,11 @@ limit_upload_speed = 100
 chain_id = 'testchain_1-1'
 token = 'ustos'
 stratos_chain_url = 'http://127.0.0.1:1317'
-rest_port = '18091'
-internal_port = ''
-rpc_port = '8135'
-monitor_port = '5678'
-metrics_port = '8765'
+rest_port = '9201'
+internal_port = '9301'
+rpc_port = '9401'
+monitor_port = '9501'
+metrics_port = '9601'
 traffic_log_interval = 10
 max_connection = 1000
 [version]
