+++ example/network/node10/accounts/st1cyl5vsn7szkfz20ulqcej43heu0q3cdtwlsmuw.json
@@ -1,0 +1,1 @@
+{"address":"c13f46427e80ac9129fcf831995637cf1e08e1ab","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"24793bb9de55c88c8f5177119dcb062dc7e884558b543e2cf00e4228a85a518322651127f06649b1ae7c084033a3941d5927f1f474315ec4e0352a951fccc3474f584ffff8b542e836646c0493d1c8df273eaeeb074394dec88e6f3416f8530f6305bd14b1f39757c2082d08c2e05617d06ae652427308c16af3803584cca01bb93af05f4cd81530824406f1a8cd8cc1efbd50315bd2c4063e932e2d635adea516a05b3b540538a5c3a88270e30474d1a96ec771b318097be550de1fb45ac807a04c1fb5267cbdec05ea5bc01511c648be1a8afe444aa1deae15329f3c7d4e8d30a977a39ba6d2aa4cfc10d85b18a1cbbff64deb1311fa9e8e","cipherparams":{"iv":"1857986f5d11dbe07340ede5ebfeb8e4"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"14242336e8ee13dff2b376254d29f7539ec55ba5f02f339aa1bee09fc10170c9"},"mac":"cb7a179e17e7a7481ae39de0f0dc9181185796b71cef088280e18abad87847b1"},"id":"5937d260-fa8c-4bf6-9cd4-f1cfeb9fe44f","version":3}
