+++ example/network/node10/accounts/stsds18642sj09exj5vze35m9l7uhnlmjqu84dfs57jg.json
@@ -1,0 +1,1 @@
+{"address":"3eaaa849e5c9a5460b31a6cbff72f3fee40e1ead","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"3396d03fabceeb3629a99ce6ea386025ffc32dd87bbf370897d8b89d2754f877d460255eeb21ad43b7ebf9ec2dff0d083aba6e561ea97a29534c9ae06e202400e78c5d94ac7d071d023557f8ef04edf2115c3412ae715ff2b4840a1e7a255991971c232ee489a7b8b4ae79e7","cipherparams":{"iv":"e4533389eb5c931e63dae6290a984def"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"1e8a36c9ef1a935d5c12f468f8ad06d4ad823ea84764fed725bb2e1ee2594cee"},"mac":"66ee762a2cf4c23ef0f2c5b78fbcc2f30adcf729dad3d4c86a171c65b620aa99"},"id":"2a9ee8a6-4e7e-423e-bf8a-ee4dd415490f","version":3}
