+++ example/network/node11/accounts/st193jmh68aj7ggzlkmjxn0qnkj2qld093ser9etk.json
@@ -1,0 +1,1 @@
+{"address":"2c65bbe8fd9790817edb91a6f04ed2503ed79630","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"f3c3def63ee61fa59a033cde9c4b83c4499631b99351b171846cc767e456d6e4321c846ab0490af3dfbbeb4e41993984c3da306ac7af84e3b1c05eb9bdbd1de23c9df3c7408c71ecdc4560401b959fe50189eab1165b159d8a3019b8a95bfd4d97001dbc77696e9a18e8bb1e322d64bf09c39aae9ed6a49ef8242555cd6920ac2b6de7bd81ba1ba67e1f27209438105d377b331e8c47ce39e40f14051204ab5100ced633628d6b858a3643bd33529cbe17ca0a74c141a628b5921cf44c8a356b3eb189b82ba383dcbf736715bf97837c4f9c2ad55c5a5f218f544602ad9b4c7e732c2992afc9d1c7512f9ac43056dc0e79e2e5d7c1169e1e7a6a4f8e7322","cipherparams":{"iv":"620d4a48663055827b0c35b7f86bd3cf"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"c084d26e2c51269af0d737281bcbc8b51eb49131e9772717c9d3d338a17233cb"},"mac":"f889bee0e8ee908c4151a1149a29399a154a3d839dfbf6aac207b256def49ebc"},"id":"c8fe6644-f5d6-4c67-8df1-68231160c14e","version":3}
