+++ example/network/node11/accounts/stsds1y04j98lh9ht0cc4wtn6v5g5673urrgnhvdylvm.json
@@ -1,0 +1,1 @@
+{"address":"23eb229ff72dd6fc62ae5cf4ca229af47831a277","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"032db209f7e0322dabdeac6b136af531d2f6a121e07f66faa4d23e1fb807b756c2f5b3eca5eeffc41549ebc7f64e04634253b0a0a3e7cb195e696149ac73a773da2e4403bae220bc043065d6e0ec1e34a9eb51e6c640a764bff92dc8ba5946bfda7056ffee7aba38b30cf187","cipherparams":{"iv":"74f8870e14109635c5865f556a888cc6"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"5693bede28cfbd64594a61ae832fd2bd7068cbf31e0e92225decb2f029b8e77e"},"mac":"6290d285eb607f674b8528bd0b248d5478feb5763d45bb118d8ef933e79becf2"},"id":"6c8b2e9a-71a3-4502-a067-7efcc9dc5e22","version":3}
