+++ example/network/node12/accounts/st1fl69wq28dw3p2rd8mznnh3uggf44euk4p3tkwp.json
@@ -1,0 +1,1 @@
+{"address":"4ff45701476ba2150da7d8a73bc788426b5cf2d5","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"7ad16070e609abc99c088ec1f618452015ea1832c50d6bde428d1ed1d76fff021da6d6b1a4342bcd9a55b46c074bb068b417428430fdc34bb0a4208cb13c7063fdf6f73d36d3fcf875bd90c72d743b802d4a60342878b293cba6d4b80416c9ba9d1fb719567f8617b84ada2f621aab486a55d1269ab6e77d7cd78482f46e8964023fe5318400863c4d89edfe442cbb46fa849ec852567561e9eb784bc28c4b947bcb0564001510e1b27d079af400d518944cf98d312d9dd6b5b27e80518fb1673b3343eb5caed5fc0469baba22846571cb96493b192087446eb3252fa0c75fae3be18618fb6782e4b0e1d5799107d8032dc0b339","cipherparams":{"iv":"daf6fa56c6bd34d9916e57110f6d2806"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"6676b81326c3fc1c373dd849912d4b5767db83c69fc48ea8a13855906d9fd87b"},"mac":"2994f23bca036a431c9726b2af01a08da0091c7e04b5d4cc754ede1c12deae19"},"id":"836ca10e-e484-47c1-a16d-a1b819fda7e8","version":3}
