+++ example/network/node12/accounts/stsds1mrvy0p5ggf6wjwlxtewt9ehj47ctsp090ugreg.json
@@ -1,0 +1,1 @@
+{"address":"d8d84786884274e93be65e5cb2e6f2afb0b805e5","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"9cfe163bc43e001b8a049e6487fc3e71e6eb7b51dab57897ca555f2c559a197efa51ccb1a4e0603dfac10050542340c09f81d6eeae106d4f1d44596616b708be4445271a145f7016d4cecfda5a3136da690b399809fa253d7e656bc9f72629ee8ea6e187b0e0aefd77ce5975","cipherparams":{"iv":"40cb1ad7d9b6b57248614239b028a40f"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"b250708ad7b2a317a76634b6af9c4916fc0f611bf6496bc9a4e2dede9a2386e5"},"mac":"113fc75bc5b5239d917e633e9c929845c3cc205a2bfed8125ae8d19465660eb5"},"id":"c15b0308-c88a-46b5-bbd8-be2d72230e2a","version":3}
