+++ example/network/node13/accounts/st1cl83c3ght3dzzn39pnl3mct0mrlt55jztuwkf7.json
@@ -1,0 +1,1 @@
+{"address":"c7cf1c45175c5a214e250cff1de16fd8feba5242","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"e9af00fd3bbe1d9f2200c9f988704c5034ea4b257bc052dbbee77ff691d66e7aa8244adbc75f413cdc77089728f1c8615851f5538817b0ac596c75f19bbd1ed0c786b9d4d05228844d069219783a5530f3babe489d0b5014e6cf909fe01f0feff011ae81f7894967cac374dff9a090ddfdfe890d85c7ea1aa3bc43293db862ef5c4c5224c025081bf4c4430c9ed53a422323c0bc29e4f211a3f7ee7ccb7868bc9d05fce3a740d49bf07ed280a89ad78ce1f14e9902c2948a90e487749bebb0e691ceffd3e230680add7801251043b4cb68e0f716ab3b526c702aa1ab4f35f0b34a3f08e7b204fd4c198d6a0a3fe96cd09eb10d8cd56a","cipherparams":{"iv":"0a691f8e020fb8475b4480a0ea31df04"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"b92b4a9a58cd8016d13a44f2578b887105ffd9d13b2e8b9020f8ed3b6c4cbe9c"},"mac":"dafb688769dc90e526f6a70feaea4dfafea1b77b976ff9e9ea40c67b0271f901"},"id":"e4a99772-f16c-463e-b396-6e4bae397c85","version":3}
