+++ example/network/node13/accounts/stsds1h0hnfnfcv337jzyeukzxe7e2ypm0kfvtzwrxr6.json
@@ -1,0 +1,1 @@
+{"address":"bbef34cd386463e90899e5846cfb2a2076fb258b","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"ed80a4ef6edf56ebba6bfa697b87b0b57f84607c7defed807e0f1ea0a5d3caa4eb5dd921955fb52b962082463454a8e65322cd8bb95784cbb1aa6bfff100c327d058fd0bd6881e0395361dea72e2c722d8c856606565c725cb3cc1297ed0796f14a0e827aac8469793f62172","cipherparams":{"iv":"5214ac55569f28274f6afd92172cdf51"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"124b8a3e2b0ea7edbcd919765d7d9360b6a4d036373a836e279a3b0521762bfa"},"mac":"61b56660f78b25c9e1b1b6d3541064c65abdb758257d9844ac59b3dfa73cb6e4"},"id":"e741e8d0-c580-49d5-acae-7bf01d1e1864","version":3}
