+++ example/network/node14/accounts/st1tg73863w42wa86m0jnhxs07fewdwdjp2ruznze.json
@@ -1,0 +1,1 @@
+{"address":"5a3d13ea2eaa9dd3eb6f94ee683fc9cb9ae6c82a","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"24bcf67bf75f5f4e291f9cfa5f0a2f0eb9018051171b732264cdfb347001479692c63d203fc485b5dd2e21dfa85727a0843224643172ff799c1a01db71f74ea9753acba89d83403d3886d70f2d1790be7f3b63c3f67a3236badca61183cce742995e5aa5f3a22ccc80be4b5d8c70cc043579c8c337af853c8db24a364a4aac6f8802c2102ea75065dff7a160f3df495c057286e5532208a8eab27119c62b23ceaaf634c5143b1ad8089bfff503d424fa926cb59719e4c3e55c6b73e2a56ad96911d929cc38eb03d030416e30e0e42fce992894945cd8916a37a1995a813e8ab9007ce84f1a8abcee9fbbdd4744eafe5c90764911d7056d7d9100","cipherparams":{"iv":"7eb5778f2f83d3bac8e531acbc37eb9e"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"1266fdfe0ababd86a9e69e822b030acdc24afb9af3371610d5852600fb5e63c8"},"mac":"8039ade094b40b37338d5bfb161f92ef7fc4fd2fd41fed8e5cda4199d5c1ead3"},"id":"8e695dfb-a6b7-4970-97e8-5af79fd0f12b","version":3}
