+++ example/network/node14/accounts/stsds1nuluvunfkk5prwtdp2xea7q3huuj5zly03vmnr.json
@@ -1,0 +1,1 @@
+{"address":"9f3fc67269b5a811b96d0a8d9ef811bf392a0be4","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"5fc744df9eaeda82459079df0ed44149df41734886fc71f28867629c3695d268ac157802332a9d1d9a8f15eeea94f5536f1df8e2183a8f3543e082421b693045dc5e409838da247eb9fb767ddd8e3be704b77aa592cde625c017dcc53e0b5688da0fda078c39aa98a4b1b815","cipherparams":{"iv":"d5646b8556e32f2e15c47387483cfc3c"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"55cfb813db96a122417c7407eb38bfc324b8269618f9f0a423d290259a77615c"},"mac":"95508fdd55f5f41631202c33f69623cdb9fd8e8631b067c0d6cdbad516300f47"},"id":"f203a09f-eea7-42d3-9751-60ca6b8cc953","version":3}
