+++ example/network/node15/accounts/st1mnness9mu5x7yevc6glrwqwny3pptn4kkkjewr.json
@@ -1,0 +1,1 @@
+{"address":"dce79840bbe50de26598d23e3701d3244215ceb6","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"e47a1e0e791a7ae961ef72f7a8a73a71b30f30dd21df1409c67071308ec0b461f368e155b0e39e6ff2687a0a1223383df72811a52247501716e40dc684792d98aa11ddd11028a00ca0a0bca9f0d13adf6beac3e6b3c8521e78e674b4a4986cad003ac42f59f4a02f98a7a5671a20a6c924c4114d84ee310ceda4db1cbafd4618be0923922d745b69fb9bda55224769a61fea4ccf74e23c7fb9b4d290bc72a286bfa08b10627d7542d30b85f7a6bb8b930a2839835012552be9cd046cd5125cb8cbf54e98021816f0d02637888f8156fbc64a0bbfb9e5f1d319dce8c00445576cc45eac99c8a543557abc1076493381b6f293186b","cipherparams":{"iv":"b00d324b816bd5b77f20eb1235177576"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"51b894458a392c8867879d3b48024328bcd6c2a6c5f4f133e141f84fb1a49dcd"},"mac":"562ad58fff32ee094f4ae687739c156035cbcdc264af166a794faa82fcf59c3e"},"id":"1647a2cb-7ee4-4440-8da6-b1404849b297","version":3}
