+++ example/network/node15/accounts/stsds1ke6juwvnglgvj67kdvkruse4rqcln4wragk0q9.json
@@ -1,0 +1,1 @@
+{"address":"b6752e399347d0c96bd66b2c3e43351831f9d5c3","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"900a27f4cf3f797f94c9742ebe8a6f27f363cb49fef191e85d5f3ba384112dffb1e696c1c862e8c5f8e16c5a3202b1b2f024ef2c8bff1ceb36ab28a40a7f086959c3ac29bac05a830ddc0e5a84919c6d654e90be6e0d50f03b1ec3c34d60d11ee10e7ac23fb2553617b70fef","cipherparams":{"iv":"6aedc79e54d6fd13a277d39cd58972f1"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"0679ddf3d6069120c7009582fd9f7eb3452cfdbe623030a1a4589c890eb1c17c"},"mac":"39113fbbfdb1f9aa5414db1f9f188263d95b69c884e303eb478f5c8cc88a1809"},"id":"4cf413f8-931a-4ceb-be29-d4d8614710ba","version":3}
