+++ example/network/node16/accounts/st18dqqsa86w093m6d79zgrwymkqnj0nel6fp3ztl.json
@@ -1,0 +1,1 @@
+{"address":"3b400874fa73cb1de9be289037137604e4f9e7fa","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"42d48bd6e9208edef84a8d1899ff0d7e5ef6c41ed5e5eed7d5bf9b44564e1c6709ae355ef5302917deb5e264ab2098579f65cd5319d9a6711a8129b5268a57d8823f51f4063df61198f23e01e39c3847c2d8ad1c8465bb32e0ab9816d0980c8d026f55b56de469dc0354e88fa62a77d41d47911452cd653c7b90ee1e497aefe9a83229b0bf116c0e7b59df05ca581241d2c7acb26c785509f2e8a74c13f4d2e9c85403938c025e0eceb3bf775ac3d3847e209145060e29528c7a8f46dd2addbecc09880bcca8b4ee9571a13bab1cb95ed2936ab12ea9562fb0010220ca93f767607f98c530d3b80a583bd3a00c4dc9b11372b7","cipherparams":{"iv":"1671f69c10f13b8d3b9dc29647a4e81d"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"f67f7dc653ed4aaa228ca2039f68df7fc0b09736ca179bb68a826afbf5e22c10"},"mac":"94bb624504dd1541a78151da13d0ea1670d2ac555c6565e03bc6033634fd4f69"},"id":"6073762c-b5f1-4928-8884-e1c2c320ed92","version":3}
