+++ example/network/node16/accounts/stsds1qr6wdm5nl2kzqm3r95gkzkw0kt0wjx3lsqpzac.json
@@ -1,0 +1,1 @@
+{"address":"00f4e6ee93faac206e232d116159cfb2dee91a3f","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"dd03bda3f5e5203e3abe0ca17a18758155ceb763912f029b190d6d7c6c09be3aa199ef417d21883987d4dc9405cade0caa8c0d82b8edae1b2070d46f88574033387053eaa1113e95cfb6e19ae86ff4ae62d99bbd720b2c778b09f1f0ff073ec17d34a564d6aba99b82e1a03e","cipherparams":{"iv":"279f5c9487edb3c23236ed3a0f35936f"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"ce99595a0632842048146b673e0c112ff122770df045374683b63c3ee21f5916"},"mac":"e5e40395a428f9134971e2d5e66c845edbc96f68eb19f64f51466cbe390a3276"},"id":"30a7a987-072a-4671-8e77-2b24d63abc89","version":3}
