+++ example/network/node5/accounts/st17qnaef7vkkkx5vqlju2amsu7xatxfwsyu7l2jw.json
@@ -1,0 +1,1 @@
+{"address":"f027dca7ccb5ac6a301f9715ddc39e375664ba04","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"2fb778f68ccb904d6fd347a4a4ff6e203ece3ae259d185cb13747735aedf9843e6ed1f86f1dfa9503bd95d39c76eb20321277f8ee85c4dacf6ef1efc660c28f6aa9530a241e9574eac5546d384fc3df5991388dbd8a335ed9086d4f6b5b81687ef7858dd96c2fa60dfc034c1df7c6aad58bfac6b12941f6f6a7147e7e41c6112d785c686def5f04cf18a8966ce20ab4b6448c173184a27668372a6f4dc33dc86af80629e38dca5baaed1f508b863a99ea1bd1fd230882b9e3feec6f490ece7dcfd8895c008749277a359a88b9190557997359492edea80fe7338b5582ae6f1a313cc2ed11d061994467d52bf1fbabb8021ec7876d9724165","cipherparams":{"iv":"4b1503b41c59250fa8197105b4fda2d0"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"533dcca6353fb06e6f445890d2713abe638e2af68781db81fb2a5f8e090c4785"},"mac":"df15547b356f35a40f04d24385438e88a8150554d88c66fc51654d7b7716be6a"},"id":"7712a7b1-a64a-44fa-aeab-b36959057654","version":3}
