+++ example/network/node5/accounts/stsds1gg584kpjuhqw3re8hf4hwj3qwaeyvjjvfl8387.json
@@ -1,0 +1,1 @@
+{"address":"42287ad832e5c0e88f27ba6b774a207772464a4c","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"066dc397d691639f196581525e3a66f2d533a3ebc92a474bbcda23d5ce1c7c75336c330d21c11e454308af97fd2f3e7083a2cbb8172b2e48a71058e71255d2435debcf2f6c8abadbef008887aeb2d723ca29e17345bba12681f8d2dd2e23a763de5aba6317388c44bd519410","cipherparams":{"iv":"db324607a80832183eed867fbc154dec"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"7b192cc5994fdb23506b37db0a0697ea164e49f4c45e37d132ee67c4fb0b89d4"},"mac":"b800c58e8655dbe085a341154ff0d982a53e76de1d4c8c2a60edd374541fe8ff"},"id":"dc26880b-dd5a-45a0-ac50-dad0e597b09b","version":3}
