+++ example/network/node6/accounts/st1ctdqs7xvycdxkjaggfw3k4wlv4k2sndvfh5qmm.json
@@ -1,0 +1,1 @@
+{"address":"c2da0878cc261a6b4ba8425d1b55df656ca84dac","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"50cf4a421da0bab04cd4a87f5d836e33b71e75f11cd617ef7671f9f934dbaf2a0a83bfe69adbf7849a0ee2357833d61623ad2ab22d4b8598b0e7a33f023578c7865d5271745936b8702102f9f936190c84dd03da835273983d0f0bb440ee0c25c302b5cb6e8f1acbb76cc5ba0eef77b7247a149ca4a4e27d4fa74711ccc4e17989d9b0da642e4d5a36c7dbed10d5a8302bf19733ac05004fed1be8834aac308917f5a04b9d68b27c8eadd3f92299db06b6b624ef9fd0b45c7e8c68d88ccc33276f4662d8eda78ef81831ffe0b5d4fe74f4ffa3cfd15812bdb6c8df0033a7bbe0f7e32b242c339f382539c19c3c78010162","cipherparams":{"iv":"c53959f6007058b9ddf68fe385efd996"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"062e8e8b50816260d3712c7b026cdeb36d26a6b6db2305c2b6bdbdc07149d4a9"},"mac":"212f954040aaad31befc41ce9cdcaa66766f75b597ca6283a1a3a6d2a3ce06a4"},"id":"8cc4776c-6316-4a54-99fa-da1538848233","version":3}
