+++ example/network/node6/accounts/stsds1ynz9z2yr567uatzqjuwj6ll0ruq5qf7q4m2ga8.json
@@ -1,0 +1,1 @@
+{"address":"24c4512883a6bdceac40971d2d7fef1f014027c0","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"b8a22f0f2e31e995e2674392512186814485c8fc98a215e538ac025ed2f72b2838be4aab53aaddaedaed83c517caac227891244803cab00c19f89f92100d5b897eb436842f7b732798d99c94b00c351bb60c0880d4ced62972f5edc58980a931fa4ced8d10160e93cca01e95","cipherparams":{"iv":"74ed553ef7234487b72e2edcbcfe87e3"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"762e62a150bc9cc21ca5e62e747ef5a2d2fc28fb6ea9bf5f233befcae93b2534"},"mac":"4a880642a05407b695ccbdd85e46c60b041b3f20a327d4fcea5304b7d5f98dc5"},"id":"d26d799e-d9b4-4648-9939-4f8b9b287e2f","version":3}
