+++ example/network/node7/accounts/st1xx9lqrrsuhdkzsgkt88fywl5kax2vzrp87qgev.json
@@ -1,0 +1,1 @@
+{"address":"318bf00c70e5db61411659ce923bf4b74ca60861","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"b3bddf71a0f6a5bc3ee12371686b5f8729a7678d7c709a49e778486ab755548740ddae3f3ecb3fb2d5662b06a81cde81d5cbce4c8e8e14cdbf7d52fa7dc963b2b517605cc58e896667fc36892b7e37f0e008d6b5ecb0fcc16587a0aac2856cfa3adb062a8f4501c6a05e300eb7831ca3e99e176d05a7e6ea50367db7b35e9bdfa22a8fd7162dae3c4e8c5f9ab83df7d55a9f7363db0cc64a81a9d831c33e024665ed305d1cd76fe4d08f533ac9ea3fdad96f21b956b656d3ef728a4792de2c6ad75a5792013b2fc545078c36f803c6ca18bc67d0e3486a83a66dad3d1ccb93c538122da77b8f93b2ebe2c5d223d37657eaf08a38c060f76361f8d6a8841b","cipherparams":{"iv":"c01beff159c771c43956c0b37bb7a5aa"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"0073125893f53437aa8ca81000e4245a5054f38ea6c0215715ee9a9508602fb8"},"mac":"7bcb0c789edea1d075fa45f9f79d7b617b5680955815c2efd479e4fc23c8bc73"},"id":"95210fd7-bd60-4f9f-b237-9716a252b0bb","version":3}
