+++ example/network/node7/accounts/stsds1td7zr0eup374ahrt3wjr08gmjueadxuan0eewe.json
@@ -1,0 +1,1 @@
+{"address":"5b7c21bf3c0c7d5edc6b8ba4379d1b9733d69b9d","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"78fc76aa59567eb1c70fe945a07c8bc1339c77becefdebe680cfc7bd9e05e2d9d3dbf00ddf1432c7040789e7c4d0e1d67b75dc8275c23bf5cfef4c7d7f787636d0149edbdc35a5b5308ee3a67a328eafadcb5aca27f8d06043faa016d934af343b6d68aca07de09d7f52ab14","cipherparams":{"iv":"0b29bcdf83a2db493962189b31d1293e"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"caa23b2b9e041732b2a109f6dbc0c00335403ecdbece307ef6ed0a1332bbbc84"},"mac":"32bc189ca92706401893c2bde188ddf53c9a18779e1ec5effd9c25ed11a1754f"},"id":"548e412e-f345-4f3d-b841-ce4ef7a19272","version":3}
