+++ example/network/node8/accounts/st1d2jnpmg3f29ysym7nfg9ln3mphqdvl0turnq7h.json
@@ -1,0 +1,1 @@
+{"address":"6aa530ed114a8a48137e9a505fce3b0dc0d67deb","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"98dcd25359a4a97c3b116f6b149c9e6e71dc8c08a694ccb55b1a5f015f2fcdfdf0933e2835097138bdc6694eef43e59aaf87726ec9af568c8bab81f4f1702ce30179bd4eea02ccdff6bafc3173a187cc48bf1e0a17cbf85e8237da77a35e9d099832568ea6e2d77d52294d38256d313f1963e743420b9861aa955be9812ee7f3a32eb1d4c42a69cbb9eb094adf195496e718b7fed71bb8d01d285c68934b388a515a10f19e0394fba5d73770b16d64f8c60bc35d0f9e273d9d9482becf5963bccbd3a60ca817069c060d317402c8290e89ff06ae7af62036a82e0b5736948150b8b47873bad6be50b0fbe80f34099a0e1ee1b1d8ea21","cipherparams":{"iv":"17b7d99bab8073537797b8535ecdbc36"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"c1449cff967c5de63532d05d39f5be0be0ed0186496540ab380d0b9c56ae3539"},"mac":"c16b74845c923cb8fd72e258cbf0b09156ea9ce63bb9475f80a144e3c6c39491"},"id":"77a6d2c6-7290-4dc1-8d5b-1131a3c39114","version":3}
