+++ example/network/node8/accounts/stsds1ggvr68mwhe3tcuwa222qnzh665c3cvvnf2kksk.json
@@ -1,0 +1,1 @@
+{"address":"42183d1f6ebe62bc71dd5294098afad5311c3193","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"5f89bf403420ecfb6e5e60214b5caf904c205ef9452454da6dae06f69619549432ecfe2da5b16340c8d3bfd6627fc73d8ba388858aecc6fd3767f6cbe5221f54ed0c18f16580afc8ea056f32332c8026b338dbf7094c13f315dff1a8b1c2306c38670253eb5bc59fe6f55ab6","cipherparams":{"iv":"4a74ad7baabdb63b1f702d4ceb17e493"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"3d5346be5a7c8ca66d7af67f4f2b6c8d48878141b2c17f488e902d3f72a6fbd0"},"mac":"a5999f8e74219907366d847ebfde44f5ea7dd5b88232216006f2d88238181944"},"id":"12d3335f-0fc1-4156-bb4d-7d8fb7f8e1f2","version":3}
