+++ example/network/node9/accounts/st1fne8snamxwa39hd5ajtvnscfwdpap8k99gcg6k.json
@@ -1,0 +1,1 @@
+{"address":"4cf2784fbb33bb12ddb4ec96c9c3097343d09ec5","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"62310e6078d16e8037a02f333ff8363f74faf1ea3b2e44f421b7ce67ef67042cfb42b3dc557e55565aec526f2b2a100a6247d0c30e0a36d226054067b7ee09715f30dd98425c76889f3086753e3f037971604d4d1a2c4289093f97c766c665a260195fdb4fd5531aa4a30ce9b499721cf02f7e384a5d67b8e8c0466fbbfbacec86d324bb352c9c278bee12d6376dc418d516c81a145e009f48e819028423f438cc12626b0d84a2f24d306b5eaa63c725016833323fe8720b292b2a185029e348e3f1bd31ad5bbfa73485f9e141452dcf7b4ade5699b140c1303aa3e3f907d83a40ccf3e652f7d7ed0cd5e55e10030850a903bcec541e103fae9b15062c03526b","cipherparams":{"iv":"26d07959da0e60b8447383e9fd7fbac4"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"1f31c0fee6afb8898b3c93276ae3b628f6a9f7419836b36ceacd531ee6983232"},"mac":"e28e1fbbe7163142d27a0999d162ed3e68556f48a85daf3ecc6814ec02314e92"},"id":"742abc55-d8c7-4d57-8a99-8d22c436715d","version":3}
