+++ example/network/node9/accounts/stsds1lp9kn9js79q8lutzyuzrsj8zaju7n3y503rt05.json
@@ -1,0 +1,1 @@
+{"address":"f84b699650f1407ff16227043848e2ecb9e9c494","name":"","crypto":{"cipher":"aes-128-ctr","ciphertext":"903278aaefe41905e3de7aba67a79c2107a681c24d51168649ba7bf34a0c81e8f21283247aec2ce93539aab8fab70d331b30456741e68c9530f1b212c94951b823e4d39b0f4db2245ddaa07b6574abae184a6ef184b7cc9d2530a1ac86f4ad18dcd7d4fef3526a4d77fe3c78","cipherparams":{"iv":"73c8ee295b9253bd70a22d89d4c5a08b"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":4096,"p":6,"r":8,"salt":"aca41687e1bcbbb79eb23e3b1f68bc28845503410cd36c265c512fad580cd9a8"},"mac":"785ff3cb3328a6c7312d2daa5782f6caa8bc46ed41ab72f1e1a9210bd3ac30eb"},"id":"64ca3678-4f84-4f93-9ce3-4f8a4cb57d10","version":3}
