+++ pp/task/transfer_task.go
@@ -9,7 +9,7 @@ import (
 )
 
 type TransferTask struct {
-	FromSp           bool
+	IsReceiver       bool
 	DeleteOrigin     bool
 	PpInfo           *protos.PPBaseInfo
 	SliceStorageInfo *protos.SliceStorageInfo
@@ -19,7 +19,6 @@ type TransferTask struct {
 
 var rwmutex sync.RWMutex
 
-// transferTaskMap
 var transferTaskMap = make(map[string]TransferTask)
 
 // CheckTransfer check whether can transfer
@@ -36,9 +35,9 @@ func AddTransferTask(taskId, sliceHash string, tTask TransferTask) {
 }
 
 func GetTransferTask(taskId, sliceHash string) (tTask TransferTask, ok bool) {
-	rwmutex.Lock()
+	rwmutex.RLock()
 	tTask, ok = transferTaskMap[taskId+sliceHash]
-	rwmutex.Unlock()
+	rwmutex.RUnlock()
 	return
 }
 
@@ -48,7 +47,6 @@ func CleanTransferTask(taskId, sliceHash string) {
 	rwmutex.Unlock()
 }
 
-// GetTransferSliceData
 func GetTransferSliceData(taskId, sliceHash string) []byte {
 	if tTask, ok := GetTransferTask(taskId, sliceHash); ok {
 		data := file.GetSliceData(tTask.SliceStorageInfo.SliceHash)
@@ -57,7 +55,6 @@ func GetTransferSliceData(taskId, sliceHash string) []byte {
 	return nil
 }
 
-// SaveTransferData
 func SaveTransferData(target *protos.RspTransferDownload) bool {
 	if tTask, ok := GetTransferTask(target.TaskId, target.SliceHash); ok {
 		save := file.SaveSliceData(target.Data, tTask.SliceStorageInfo.SliceHash, target.Offset)
